+++ internal/tests/benchmark/query_test.go
@@ -35,7 +35,7 @@ func getBenchmarkQueryBulkLoneMultiNsPrimitives(b *testing.B) (benchFunc func(),
 	assert.NoError(b, err)
 	bytes, err := internal.RecordToBytes(record)
 	assert.NoError(b, err)
-	tf, err := fixtures.NewTestFixture(&fixtures.MockServerConfig{
+	tf, err := NewTestFixture(&fixtures.MockServerConfig{
 		QueryBulkResponse: &commonv1.OnlineQueryBulkResponse{
 			ScalarsData: bytes,
 		},
